Add optional data interval to CronTriggerTimetable
This allows the user to provide a (static) data interval for DAG runs to
adopt. This should be useful for most cron usages and more natural to
the old schedule_interval logic.

Also moved CronTriggerTimetable to its own module. The class living in
'interval' is weird.
