+++ tests/timetables/test_interval_timetable.py
@@ -26,11 +26,7 @@ import pytest
 from airflow.exceptions import AirflowTimetableInvalid
 from airflow.settings import TIMEZONE
 from airflow.timetables.base import DagRunInfo, DataInterval, TimeRestriction, Timetable
-from airflow.timetables.interval import (
-    CronDataIntervalTimetable,
-    CronTriggerTimetable,
-    DeltaDataIntervalTimetable,
-)
+from airflow.timetables.interval import CronDataIntervalTimetable, DeltaDataIntervalTimetable
 
 START_DATE = pendulum.DateTime(2021, 9, 4, tzinfo=TIMEZONE)
 
@@ -43,12 +39,10 @@ CURRENT_TIME = pendulum.DateTime(2021, 9, 7, tzinfo=TIMEZONE)
 YESTERDAY = CURRENT_TIME - datetime.timedelta(days=1)
 
 HOURLY_CRON_TIMETABLE = CronDataIntervalTimetable("@hourly", TIMEZONE)
-HOURLY_CRON_TRIGGER_TIMETABLE = CronTriggerTimetable("@hourly", TIMEZONE)
 HOURLY_TIMEDELTA_TIMETABLE = DeltaDataIntervalTimetable(datetime.timedelta(hours=1))
 HOURLY_RELATIVEDELTA_TIMETABLE = DeltaDataIntervalTimetable(dateutil.relativedelta.relativedelta(hours=1))
 
 CRON_TIMETABLE = CronDataIntervalTimetable("30 16 * * *", TIMEZONE)
-CRON_TRIGGER_TIMETABLE = CronTriggerTimetable("30 16 * * *", TIMEZONE)
 DELTA_FROM_MIDNIGHT = datetime.timedelta(minutes=30, hours=16)
 
 
@@ -67,23 +61,6 @@ def test_no_catchup_first_starts_at_current_time(
     )
     expected_start = YESTERDAY + DELTA_FROM_MIDNIGHT
     assert next_info == DagRunInfo.interval(start=expected_start, end=CURRENT_TIME + DELTA_FROM_MIDNIGHT)
-
-
-@pytest.mark.parametrize(
-    "last_automated_data_interval",
-    [pytest.param(None, id="first-run"), pytest.param(PREV_DATA_INTERVAL_EXACT, id="subsequent")],
-)
-@freezegun.freeze_time(CURRENT_TIME)
-def test_daily_cron_trigger_no_catchup_first_starts_at_next_schedule(
-    last_automated_data_interval: Optional[DataInterval],
-) -> None:
-    """If ``catchup=False`` and start_date is a day before"""
-    next_info = CRON_TRIGGER_TIMETABLE.next_dagrun_info(
-        last_automated_data_interval=last_automated_data_interval,
-        restriction=TimeRestriction(earliest=YESTERDAY, latest=None, catchup=False),
-    )
-    expected = CURRENT_TIME + DELTA_FROM_MIDNIGHT
-    assert next_info == DagRunInfo.exact(expected)
 
 
 @pytest.mark.parametrize(
@@ -131,87 +108,9 @@ def test_catchup_next_info_starts_at_previous_interval_end(timetable: Timetable)
 
 
 @pytest.mark.parametrize(
-    "current_time, earliest, expected",
-    [
-        pytest.param(
-            pendulum.DateTime(2022, 7, 27, 0, 0, 0, tzinfo=TIMEZONE),
-            START_DATE,
-            DagRunInfo.exact(pendulum.DateTime(2022, 7, 27, 0, 0, 0, tzinfo=TIMEZONE)),
-            id="current_time_on_boundary",
-        ),
-        pytest.param(
-            pendulum.DateTime(2022, 7, 27, 0, 30, 0, tzinfo=TIMEZONE),
-            START_DATE,
-            DagRunInfo.exact(pendulum.DateTime(2022, 7, 27, 1, 0, 0, tzinfo=TIMEZONE)),
-            id="current_time_not_on_boundary",
-        ),
-        pytest.param(
-            pendulum.DateTime(2022, 7, 27, 0, 30, 0, tzinfo=TIMEZONE),
-            pendulum.DateTime(2199, 12, 31, 22, 30, 0, tzinfo=TIMEZONE),
-            DagRunInfo.exact(pendulum.DateTime(2199, 12, 31, 23, 0, 0, tzinfo=TIMEZONE)),
-            id="future_start_date",
-        ),
-    ],
-)
-def test_hourly_cron_trigger_no_catchup_next_info(
-    current_time: pendulum.DateTime,
-    earliest: pendulum.DateTime,
-    expected: DagRunInfo,
-) -> None:
-    with freezegun.freeze_time(current_time):
-        next_info = HOURLY_CRON_TRIGGER_TIMETABLE.next_dagrun_info(
-            last_automated_data_interval=PREV_DATA_INTERVAL_EXACT,
-            restriction=TimeRestriction(earliest=earliest, latest=None, catchup=False),
-        )
-    assert next_info == expected
-
-
-@pytest.mark.parametrize(
-    "last_automated_data_interval, earliest, expected",
-    [
-        pytest.param(
-            DataInterval.exact(pendulum.DateTime(2022, 7, 27, 0, 0, 0, tzinfo=TIMEZONE)),
-            START_DATE,
-            DagRunInfo.exact(pendulum.DateTime(2022, 7, 27, 1, 0, 0, tzinfo=TIMEZONE)),
-            id="last_automated_on_boundary",
-        ),
-        pytest.param(
-            DataInterval.exact(pendulum.DateTime(2022, 7, 27, 0, 30, 0, tzinfo=TIMEZONE)),
-            START_DATE,
-            DagRunInfo.exact(pendulum.DateTime(2022, 7, 27, 1, 0, 0, tzinfo=TIMEZONE)),
-            id="last_automated_not_on_boundary",
-        ),
-        pytest.param(
-            None,
-            pendulum.DateTime(2022, 7, 27, 0, 0, 0, tzinfo=TIMEZONE),
-            DagRunInfo.exact(pendulum.DateTime(2022, 7, 27, 0, 0, 0, tzinfo=TIMEZONE)),
-            id="no_last_automated_with_earliest_on_boundary",
-        ),
-        pytest.param(
-            None,
-            pendulum.DateTime(2022, 7, 27, 0, 30, 0, tzinfo=TIMEZONE),
-            DagRunInfo.exact(pendulum.DateTime(2022, 7, 27, 1, 0, 0, tzinfo=TIMEZONE)),
-            id="no_last_automated_with_earliest_not_on_boundary",
-        ),
-    ],
-)
-def test_hourly_cron_trigger_catchup_next_info(
-    last_automated_data_interval: DataInterval,
-    earliest: pendulum.DateTime,
-    expected: DagRunInfo,
-) -> None:
-    next_info = HOURLY_CRON_TRIGGER_TIMETABLE.next_dagrun_info(
-        last_automated_data_interval=last_automated_data_interval,
-        restriction=TimeRestriction(earliest=earliest, latest=None, catchup=True),
-    )
-    assert next_info == expected
-
-
-@pytest.mark.parametrize(
     "timetable",
     [
         pytest.param(HOURLY_CRON_TIMETABLE, id="cron"),
-        pytest.param(HOURLY_CRON_TRIGGER_TIMETABLE, id="cron_trigger"),
         pytest.param(HOURLY_TIMEDELTA_TIMETABLE, id="timedelta"),
         pytest.param(HOURLY_RELATIVEDELTA_TIMETABLE, id="relativedelta"),
     ],
@@ -227,11 +126,6 @@ def test_validate_success(timetable: Timetable) -> None:
             CronDataIntervalTimetable("0 0 1 13 0", TIMEZONE),
             "[0 0 1 13 0] is not acceptable, out of range",
             id="invalid-cron",
-        ),
-        pytest.param(
-            CronTriggerTimetable("0 0 1 13 0", TIMEZONE),
-            "[0 0 1 13 0] is not acceptable, out of range",
-            id="invalid-cron-trigger",
         ),
         pytest.param(
             DeltaDataIntervalTimetable(datetime.timedelta()),
